requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.39.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
urllib3==2.4.0
